Add files via upload
The case study is to model the fuel efficiency of cars, particularly the EPA SmartWay certified
cars. The EPA SmartWay program certifies the 20% lowest-emitting passenger vehicles based
on greenhouse gas and smog ratings. For each car model, various characteristics are given. For
this case study, we are interested in predicting the fuel economy (based on MPG Ratings) of a
new car given the engine characteristic (engine size). We are also interested in further
understanding the factors influencing the green emission ratings (as represented by the Smog
Rating and Greenhouse Score), and any pertinent insights that can be extracted from the data.
The data is published by the US Department of Energy’s Office of Energy (DOE) Efficiency and
Renewable Energy and the US Environmental Protection Agency (EPA).
You will use the EPA Green Vehicle 2016 data file to develop your model (training set) and the
2017 data file to test your model (testing data).